refactor: 푸시 서비스워커 리팩토링 및 콘솔 삭제
