+++ frontend/src/push/pushStatus.ts
@@ -1,16 +1,16 @@
 type PushStatus = {
   pushSupport: boolean;
-  serviceWorkerRegistration: ServiceWorkerRegistration | null;
   pushSubscription: PushSubscription | null;
-  notificationPermission: 'granted' | 'default' | 'denied' | false;
+  serviceWorkerRegistration?: ServiceWorkerRegistration;
+  notificationPermission?: 'granted' | 'default' | 'denied';
 };
 
 export const isSupported = () =>
   'serviceWorker' in navigator && 'Notification' in window && 'PushManager' in window;
 
 export const pushStatus: PushStatus = {
-  pushSupport: isSupported(),
-  serviceWorkerRegistration: null,
+  pushSupport: false,
   pushSubscription: null,
-  notificationPermission: isSupported() && Notification.permission,
+  serviceWorkerRegistration: undefined,
+  notificationPermission: undefined,
 };
